arreglo final por las letras en blanco
se modifico caleta de cosas
