+++ Sistema_Basico/src/components/Card.tsx
@@ -6,8 +6,8 @@ type Props = {
 function Card({ title, value }: Props) {
   return (
     <div className="bg-gray-100 text-black p-4 rounded-lg shadow flex flex-col">
-      <span className="text-gray-600 text-sm">{title}</span>
-      <span className="text-2xl font-bold mt-2">{value}</span>
+      <span className="text-black text-sm">{title}</span>
+      <span className="text-black text-2xl font-bold mt-2">{value}</span>
     </div>
   );
 }
