+++ Sistema_Basico/src/components/ui/avatar.tsx
@@ -42,7 +42,7 @@ function AvatarFallback({
     <AvatarPrimitive.Fallback
       data-slot="avatar-fallback"
       className={cn(
-        "bg-muted flex size-full items-center justify-center rounded-full",
+        "bg-muted flex size-full items-center justify-center rounded-full text-black", // Añadí "text-black"
         className,
       )}
       {...props}
